added the option to finetune existing models
as the dataset gets bigger now I added a a second
layer of model training. It uses the old trained
as base and builds on top of it.
This can later be used to finetune a model for a
spezific Sample type or more general to improve
the whole model when more Data is in reach.