changed algorithm for counting Whiskers
The new version now looks at the actual defect
area within a polygon and calculates the total
amount on the sample from there.
Also changed the usage so that the GUI only calls
one script to calculate area and count. This will
make it easier to implement in a no-GUI version.